fixed setClientMaxBodySize to handle unknown size unit
