--- models/CommonExceptions.cpp
@@ -1,11 +1,0 @@
-#include <CommonExceptions.hpp>
-
-const char *CommonExceptions::OpenFileException::what() const throw()
-{
-    return "CommonExceptions::OpenFileException";
-}
-
-const char *CommonExceptions::InititalaizingException::what() const throw()
-{
-    return "Error while initializing/Checking config file";
-}
